+++ doc/python/pages/changelog.rst
@@ -150,6 +150,7 @@ Changelog
     :ref:`platform.glfw.Application.is_key_pressed()`
 -   Exposed all :ref:`platform.sdl2.Application.Configuration.WindowFlags` and
     :ref:`platform.glfw.Application.Configuration.WindowFlags`
+-   Exposed the new :ref:`primitives.CubeFlags`
 -   Exposed the new :ref:`text.AbstractShaper`, :ref:`text.RendererCore`,
     :ref:`text.Renderer`, :ref:`text.RendererGL` classes as well as the new
     :ref:`text.Feature`, :ref:`text.Script` enums and the